backend/requirements.txt
boto3==1.26.153
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart